Add files via upload
This thesis argues that neurowarfare the militarization of neuroscience and neurotechnology has already begun and represents a new domain of conflict centered on the human brain. Using unclassified sources, the authors examine the phenomenon of “Havana Syndrome” as evidence of potential neuroweapon use and situate it within the broader context of Great Power Competition (GPC) among the United States, China, and Russia.